+++ src/so-4178-6918/qs-main.c
@@ -1,8 +1,7 @@
 /* SO 4178-6918 */
 /* Common main program */
 
-#include "qs-mon.h"
-#include <stdbool.h>
+#include "qs-main.h"
 #include <stdio.h>
 #include <string.h>
 
@@ -37,4 +36,23 @@ int main(void)
     }
 
     return 0;
+}
+
+void dump_data(const char *tag, int *data, int start, int end)
+{
+    printf("%s (%d):", tag, end - start + 1);
+    for (int i = start; i <= end; i++)
+    {
+        printf(" %d", data[i]);
+    }
+    putchar('\n');
+}
+
+void check_sorted(int *data, int lo, int hi)
+{
+    for (int i = lo + 1; i < hi; i++)
+    {
+        if (data[i-1] > data[i])
+            printf("Inversion @ A[%d] = %d vs A[%d] = %d\n", i-1, data[i-1], i, data[i]);
+    }
 }
