+++ src/so-4178-6918/qs-main.h
